test(cli): extract population placement play tests
Move worker-assignment and city-expansion command/postcondition coverage out of the broad play monolith into a focused population-placement test file and wire that file into test:cli:play.

The focused suite owns the ASSIGN_WORKER and EXPAND validation/send fakes plus population-ready postcondition snapshots. The monolith keeps read-side population-placement, ready-city, and compact-priorities projection fixtures where broader tests still consume them.

Proof boundary: test ownership modularization only. No CLI runtime behavior, direct-control behavior, live-game proof, gameplay send, play-thread workflow, oRPC/Studio transport, or relationship/city-state/suzerain labeling behavior changed.

Gates: git diff --check; bun run --cwd packages/cli test -- game.play.population-placement.test.ts; bun run --cwd packages/cli test -- game.play.test.ts -t "population-placement|ready-city|compact priorities"; bun run check:cli; bun run test:cli:play.
