+++ package.json
@@ -24,7 +24,7 @@
     "test": "turbo run test",
     "test:ci": "turbo run test",
     "test:cli": "turbo run test --filter=@mateicanavra/civ7-cli",
-    "test:cli:play": "turbo run build --filter=@mateicanavra/civ7-cli && bun run --cwd packages/cli test -- game.play.test.ts game.play.end-turn.test.ts game.play.first-meet.test.ts game.play.operation-wrappers.test.ts game.play.technology.test.ts game.play.culture.test.ts game.play.narrative.test.ts game.play.diplomacy-response.test.ts game.play.production.test.ts game.play.celebration-government.test.ts",
+    "test:cli:play": "turbo run build --filter=@mateicanavra/civ7-cli && bun run --cwd packages/cli test -- game.play.test.ts game.play.end-turn.test.ts game.play.population-placement.test.ts game.play.first-meet.test.ts game.play.operation-wrappers.test.ts game.play.technology.test.ts game.play.culture.test.ts game.play.narrative.test.ts game.play.diplomacy-response.test.ts game.play.production.test.ts game.play.celebration-government.test.ts",
     "test:sdk": "turbo run test --filter=@mateicanavra/civ7-sdk",
     "test:architecture-cutover": "turbo run test:architecture-cutover --filter=mod-swooper-maps",
     "test:vitest": "vitest run",
